feat(index): Express Implementation CON-21
