+++ src/database/models/technicalServices.js
@@ -1,0 +1,18 @@
+const Sequelize = require('sequelize');
+
+const sequelize = require('../../configs/db_connection');
+
+const TechnicalServices = sequelize.define('technicalServices', {
+    id: {
+        type: Sequelize.INTEGER,
+        autoIncrement: true,
+        allowNull: false,
+        primaryKey: true
+    },
+    name: {
+        type: Sequelize.TEXT,
+        allowNull: false
+    }
+});
+
+module.exports = TechnicalServices;
