+++ src/database/models/technicalServicesHistory.js
@@ -1,6 +1,6 @@
 const Sequelize = require('sequelize');
 
-const sequelize = require('../configs/db_connection');
+const sequelize = require('../../configs/db_connection');
 const AccountStatus = require('./accountStatus');
 const SuscriptorData = require('./suscriptorData');
 const TechnicalServices = require('./technicalServices');
