+++ src/routes/login.js
@@ -4,7 +4,6 @@ import { getLogin, postLogin } from '../controllers/login.js';
 
 const router = express.Router();
 
-//localhost:1010/login
 router.get('/prueba', getLogin)
 router.post('/insertar', postLogin)
